[Future] Bovine → Shared
